Fix out-of-bounds array access in traversestack
An impedance mismatch between plain C and .NET: In ANSI C, it is valid for a pointer to
at the hypothetical element that is one past the end of the array.  (Attempting to
dereference such such a pointer yields undefined behavior, however.)

In .NET, this cannot be done; it will yield an out-of-bounds array access.

An instance of this error occurred while executing the 'literals.lua' test suite script.