Bypass broken test in 'big.lua'
big.lua defines a function called 'xxxx' that calls coroutine.yield, then
calls it -- but not as a coroutine.  This results in an error about
yielding past a C or metamethod or something.

The same error occurs on the Linux build of Lua I am using as a comparison,
so I'm disabling the test.