Fix 'files.lua' test script Win32 compatibility issue
The "GC test", in files.lua, which opens the same file a LOT of times,
would leave one open reference to the file; it would then subsequently
attempt to rename that file.

This works fine under most non-Windows environments, but under Windows,
you cannot (normally) rename a file while it is still open.