Fix bogus error message when DoFile() is used to run a script with at least one argument
luaL_loadbuffer (to load a script from a string) places the error message *below* any pushed arguments.
luaL_loadfile places the error message on to of the stack.

Thus, if we were passing any arguments to a script executed with DoFile(), we
would pop off the error message and all but the first argument, and return
that first argument as the error message.

Also, in FormatWin32Message, ensure that the returned string is properly truncated.